Switched to copying timestamped files (#1)
We now create new timestamped files with the latest domain data instead of restoring to an existing file. This ensures the app does not take any downtime during the restore operation (which takes up to 20s to complete with a bind mount, and 4 minutes without). We are also using bind mounts to speed up transfers.